+++ app/src/at-huybui/AndroidManifest.xml
@@ -17,5 +17,4 @@
             </intent-filter>
         </activity>
     </application>
-
 </manifest>
