+++ app/src/at-huybui/res/drawable/custom_edittext_on.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
+    android:shape="rectangle"
+    >
     <corners android:radius="@dimen/radius" />
     <stroke
         android:width="@dimen/edit_text_stroke_width"
